backport draft capabilities to separated servers; clean up more drafting stuff and allow central server to reject draft tables; add import/export capabilities for Cockatrice XML format; add /seval to be distinct from /eval (no broadcast); silence all remaining compiler warnings

+++ core/src/interpreter.cpp
@@ -1266,7 +1266,7 @@ Data Table::msgbox_tail(const Data& args)
     deque<string> text = box->Messages();
     deque<string>::iterator i;
     
-    if(n >= text.size())
+    if(n >= (int)text.size())
       i = text.begin();
     else
       i = text.end() - n;
